Fragment tab holders cleanup.
